Remove extraneous Babel traversal and re-crawl AST instead
As per the removed code comment, we couldn't remove even an empty
re-traversal of the AST because `cloneInputAst: true` on it was needed.
This is essentially just a deep structural clone of the AST and
does nothing special.

This was basically causing the traverse cache to be ignored, since
all references of the AST would be swapped out.

The Babel traverse cache keeps track of paths and scopes, and re-uses
them to avoid having to re-crawl the AST.
Hence, the `cloneInputAst: true` was needed to clear this cache.

An alternative to this is:

    require('@babel/core').traverse.cache.clear();

However, this was only needed because the `constantFoldingPlugin`
removes unused declarations. This was breaking our output since
the `@babel/plugin-transform-modules-commonjs` wasn't creating
reference paths to its CommonJS helpers.

This in turn is a bug in Babel. All calls to `hub.addHelper` add
declarations, but don't actually add references to the identifiers it
generates.

This means that `constantFoldingPlugin` only works properly if we update
the scope of the AST by re-crawling it, i.e.:

    programPath.scope.crawl();

Which is likely cheaper than either cloning the AST or wiping out the
entire traversal cache.